hand sample visualizer but formatting messed up
